requirements.txt
click==8.0.0
Flask==2.0.0
gunicorn==20.1.0
itsdangerous==2.0.0
Jinja2==3.0.0
MarkupSafe==2.0.0
Pillow==8.2.0
python-dateutil==2.8.1
pytz==2021.1
six==1.16.0
torch==1.8.1
torchvision==0.9.1
typing-extensions==3.10.0.0
Werkzeug==2.0.0

opencv-python~=4.5.2.52
numpy~=1.20.3